reorganize files to group similar components, update more buttons
