+++ apps/src/templates/VerticalImageResourceCardRow.jsx
@@ -1,7 +1,7 @@
 import React, {Component} from 'react';
 
 import ContentContainer from './ContentContainer';
-import ResourceCardResponsiveContainer from './studioHomepages/ResourceCardResponsiveContainer';
+import ResourceCardResponsiveContainer from './studioHomepages/resourceCards/ResourceCardResponsiveContainer';
 import shapes from './studioHomepages/shapes';
 import VerticalImageResourceCard from './VerticalImageResourceCard';
