+++ apps/src/templates/studioHomepages/TeacherResources.jsx
@@ -6,7 +6,7 @@ import i18n from '@cdo/locale';
 import styleConstants from '../../styleConstants';
 import ContentContainer from '../ContentContainer';
 
-import ImageResourceCard from './ImageResourceCard';
+import ImageResourceCard from './resourceCards/ImageResourceCard';
 
 const contentWidth = styleConstants['content-width'];
