+++ apps/src/templates/studioHomepages/resourceCards/ResourceCard.jsx
@@ -5,8 +5,8 @@ import {connect} from 'react-redux';
 import fontConstants from '@cdo/apps/fontConstants';
 import SafeMarkdown from '@cdo/apps/templates/SafeMarkdown';
 
-import color from '../../util/color';
-import Button from '../Button';
+import color from '../../../util/color';
+import Button from '../../Button';
 
 // If you want to include an image, you're probably looking for a ImageResourceCard.
 class ResourceCard extends Component {
