+++ apps/src/templates/studioHomepages/resourceCards/ResourceCardResponsiveContainer.jsx
